refactor speedy graph turn restriction context to its own package
